backend/requirements.txt
Django==3.2
djangorestframework==3.12
djoser==2.1
webcolors==1.11
django-cors-headers==3.13
psycopg2-binary==2.9
Pillow==9.0
pytest==6.2
pytest-django==4.4
pytest-pythonpath==0.7
PyYAML==6.0
python-dotenv==0.19
gunicorn==20.1
django-filter==23.1